+++ apps/controllers/SiteController.php
@@ -23,7 +23,7 @@ class SiteController extends Controller
     {
         $logs = new Site\Logs();
         if (!$logs->validate()) {
-            return $this->render('action/action_fail',['msg'=>$logs->getError()]);
+            return $this->render('action/fail',['msg'=>$logs->getError()]);
         }
         return $this->render('site/logs',['logs'=>$logs]);
     }